AI_CEO_Prod_v8/AI_CEO_Prod/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
openai==1.52.2
anthropic==0.37.1
google-generativeai==0.8.3
groq==0.11.0
requests==2.32.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.2.0
faster-whisper==1.0.3
edge-tts==6.1.18
numpy==1.26.4
docker==7.1.0
playwright==1.47.0
mss==9.0.1
pytesseract==0.3.13
Pillow==10.4.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyJWT==2.9.0
cryptography==43.0.1
google-api-python-client==2.149.0
google-auth==2.35.0
